split cron
